refactor: separate server and router
